+++ core/src/aggregate.ts
@@ -1,7 +1,7 @@
 import type { Stack } from "./Stack";
 import { makeStackReducer } from "./activity-utils/makeStackReducer";
 import type { DomainEvent } from "./event-types";
-import { filterEvents, validateEvents } from "./event-utils";
+import { validateEvents } from "./event-utils";
 import { compareBy, uniqBy } from "./utils";
 
 export function aggregate(inputEvents: DomainEvent[], now: number): Stack {
@@ -21,15 +21,13 @@ export function aggregate(inputEvents: DomainEvent[], now: number): Stack {
   /**
    * 3. Run reducer
    */
-  const initializedEvent = filterEvents(events, "Initialized")[0];
-  const initialStackState: Stack = {
+  const stackReducer = makeStackReducer({ now });
+  const stack = events.reduce(stackReducer, {
     activities: [],
     globalTransitionState: "idle",
     registeredActivities: [],
     transitionDuration: 0,
-  };
-  const stackReducer = makeStackReducer({ now, initializedEvent });
-  const stack = events.reduce(stackReducer, initialStackState);
+  });
 
   /**
    * 4. Post-process
